+++ lib/typescript/httpclient/endpoints/deleteTag.ts
@@ -1,0 +1,6 @@
+import {HttpClient} from '../client';
+
+export default HttpClient.prototype.deleteTag = async function deleteTag(id: string) {
+  await this.doFetchFromBackend('tags.delete', {id});
+  return Promise.resolve(true);
+};
