+++ lib/typescript/httpclient/endpoints/getConfig.ts
@@ -1,0 +1,7 @@
+import {ConfigPayload} from '../payload';
+import {HttpClient} from '../client';
+
+export default HttpClient.prototype.sendMessages = async function getConfig() {
+  const response: ConfigPayload = await this.doFetchFromBackend('client.config');
+  return response;
+};
