+++ lib/typescript/httpclient/endpoints/index.ts
@@ -1,0 +1,17 @@
+export * from './listChannels';
+export * from './exploreFacebookChannels';
+export * from './connectFacebookChannel';
+export * from './disconnectChannel';
+export * from './listConversations';
+export * from './getConversationInfo';
+export * from './readConversations';
+export * from './listMessages';
+export * from './listTags';
+export * from './createTag';
+export * from './updateTag';
+export * from './deleteTag';
+export * from './loginViaEmail';
+export * from './tagConversation';
+export * from './untagConversation';
+export * from './sendMessages';
+export * from './getConfig';
